requirements.txt
streamlit
requests
python-dotenv
gTTS
